Atualizações nas dependências do pom.xml e ids do banco alterados de Integer para Long

+++ src/main/java/com/pi/model/repository/RepoAvaliacaoCorporal.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.pi.model.entity.AvaliacaoCorporal;
 
 @Repository
-public interface RepoAvaliacaoCorporal extends JpaRepository<AvaliacaoCorporal, Integer> {
+public interface RepoAvaliacaoCorporal extends JpaRepository<AvaliacaoCorporal, Long> {
     
 }
