+++ src/main/java/com/pi/model/repository/RepoCliente.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.pi.model.entity.Cliente;
 
 @Repository
-public interface RepoCliente extends JpaRepository<Cliente, Integer> {
+public interface RepoCliente extends JpaRepository<Cliente, Long> {
     
 }
